backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
bcrypt

# Optional for ML / analytics later:
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2